backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
pymongo==4.10.1
python-dotenv==1.0.1
python-multipart==0.0.19
pytest==8.3.4
pytest-asyncio==0.24.0

